reorganise scheme tool
